+++ src/main/java/com/potato/ecommerce/domain/receiver/entity/ReceiverEntity.java
@@ -1,6 +1,7 @@
 package com.potato.ecommerce.domain.receiver.entity;
 
 import com.potato.ecommerce.domain.member.entity.MemberEntity;
+import com.potato.ecommerce.domain.receiver.dto.ReceiverForm;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -54,5 +55,24 @@ public class ReceiverEntity {
     @Column(nullable = false)
     private String zipcode;
 
+    public ReceiverForm createReceiverForm() {
+        return ReceiverForm.builder().name(this.name).phone(this.phone)
+            .addressName(this.addressName).city(this.city).street(this.street).zipcode(this.zipcode)
+            .detail(this.detail).build();
+    }
+
+    public void update(ReceiverForm dto) {
+        this.name = dto.getName();
+        this.phone = dto.getPhone();
+        this.addressName = dto.getAddressName();
+        this.city = dto.getCity();
+        this.street = dto.getStreet();
+        this.zipcode = dto.getZipcode();
+        this.detail = dto.getDetail();
+    }
+
+    public boolean isMemberNotMatch(Long memberId) {
+        return this.member.isNotMatchMember(memberId);
+    }
 
 }
